Path filter for PythonInterpreter.iter_candidates. (#1046)
This adds support for a path filter to allow unification of Pex
buildtime and runtime interpreter discovery. The PythonInterpreter
public API family of `iter`, `iter_candidates` and `all` is
type-annotated and tests are added for `iter_candidates`. To support
proper annotation of iter_candidates helpers, `@overload` support is
added for Python 2.7.

Work towards #1043