Remove `--use-first-matching-interpreter` misfeature (#1076)
This was a misfeature added to support a Pants use case, which is implemented another way.

This never worked very robustly:

> To be clear though, there are many corners this will not work for including mismatching abi (python2.7m vs python2.7mu) when the PEX contains platform specific wheels, etc.

If a user only wants to resolve for one interpreter, they should use `--python` or `--interpreter-constraints` with the upcoming `--python-path` flag.

We don't use a proper deprecation due to https://github.com/pantsbuild/pex/issues/584. We'll call attention to this in the release notes.