Add `--python-path` to change interpreter search paths when building a PEX (#1077)
Closes https://github.com/pantsbuild/pex/issues/1012.

Currently, PEX defaults to searching the $PATH when `--interpreter-constraint` and/or `--resolve-local-platforms` are used. (Otherwise, it uses the current interpreter or `--python`).

You can specify `PEX_PYTHON_PATH`, but this has several issues:

* Env vars should only be used for runtime, not build time. We break that separation here.
* This only works if `PEX_IGNORE_RCFILES` is not set to true. Pants must set this option, meaning Pants cannot leverage this mechanism.
* The behavior doesn't work as expected if running Pex as a Pex: https://github.com/pantsbuild/pex/issues/1075

Instead of further using `PEX_PYTHON_PATH`, we add a proper `--python-path` flag. In a followup, we can remove the `PEX_PYTHON_PATH` mechanism.