Switch from optparse to argparse. (#1083)
The impetus is to allow us to use a sphinx plugin to generate CLI documentation.

This is also an opportunity to simplify various aspects of CLI parsing.

For example: There was some overly-clever code to update the indexes
option dest from two different options (--indexes and --pypi). This
change does the more obvious and simple thing, and as a result by
default we pass in 'https://pypi.org/simple' explicitly, instead of
passing in None and letting the underlying code default to PyPI.

Also removes hand-written "Default: value" in the help strings,
since argparse can generate those automatically.