Upgrade Pex to Pip 20.3.1. (#1133)
Since Pip now natively supports passing multiple --platform options on
the download CLI, the only adjustments needed were to remove the
--headers and --cache-ttl support that was provided by a custom patch
that never made it into Pip due to security concerns as well as an
adjustment to the wheel install code due to the 2020-resolver being more
strict.

A new --resolver-version option is added to toggle between the 
legacy-resolver and the new 2020-resolver with the legacy-resolver being
used by default to allow Pex users to transition. A follow-up release 
will add a deprecation notice / warning for those still using the 
legacy-resolver.