Tighten requirements parsing.
In order to fix #899, we need to map local project requirements to built
wheels after the fact in order to learn the project name and version and
thus the full requirement needed for PexInfo metadata. In order to do
this with confidence, break up the ReqInfo type into three seperate
types to cover the three main parse cases.