Support venv relocation.
Venv scripts aren't relocatable without edits since their shebangs
record the absolute path of the venv interpreter. However, they should
work when invoked indirectly by the moved venv's interpreter. This
already worked for venv console scripts but it did not work for the
`pex` venv script, which re-execs when invoked with an unexpected
interpreter.

Add a relocation test that fails and fix the `pex` venv script to be
more flexible in its venv interpreter identification.