Guard against too long `--venv` mode shebangs.
The default path length to a `--venv` mode PEX python used in shebangs
for the venv `pex` script and venv console scripts is 102 characters
starting from the PEX_ROOT. Add to that 2 characters for '#!' and 4 for
the trailing ' -sE' we add and that leaves just 20 characters for the
path to the PEX_ROOT (e.g.: /home/jane/.pex) before these scripts start
failing on machines where the shebang length limit is 128 bytes, which
is not uncommon amongst Linux distributions.

Fix this for forseeable cases by keeping the existing cache structure,
which has a primary key path component of the pex hash which is cheap to
compute (pre-computed), convenient for debugging and ~guaranteed unique,
and introducing a "jump table" of venv dir symlinks in
PEX_ROOT/venvs/short. This jump table is computed once at venv
materialization time with a much shorter hash that is checked for
collisions and expanded as needed.

Fixes #1252