Fix module entrypoints.
Previously we did not invoke module entrypoints identically to Python
which could lead to unexpected errors using the pickle module.

Now we match Python for all cases except a PEX in zipapp mode with a
module entrypoint. This is a case that is also not handled by Python
itself since its zipapp module requires a function entrypoint; ie
mod:func.

Fixes #1018