Fix pyenv shim identification.
Previously shims were improperly cached without regard to their
activation status. Now a dedicated Pyenv class is used to test if an
interpreter binary is a pyenv shim, and, if so, what its activated
version is if any. This information is mixed into the binary cache key
which fixes cached identification information for pyenv shims.

An existing test is updated and expanded to test for this as well as
verifying the pyenv version configuration hierarchy of methods.

Fixes #1324