Fix pyenv shim identification. (#1325)
Previously shims were improperly cached without regard to their
activation status. Now a dedicated Pyenv class is used to test if an
interpreter binary is a pyenv shim, and, if so, resolve it to its
activated version if any. This fixes cache keying and also improves
bulk interpreter identification by allowing de-duping active shims and
skipping inactive shims.

An existing test is updated and expanded to test for this as well as
verifying the pyenv version configuration hierarchy of methods.

Fixes #1324