Allow --platform resolves for current interpreter.
Previously, if a `--platform` resolve was being executed with an
interpreter that matched the platform, a regular full-featured
interpreter resolve was performed. This prevented, for example,
resolving using a CPython 3.8 interpreter on a manylinux2014 capable
host for deployment to a host that is only manylinux2010 capable. The
`--resolve-local-platforms` option exists to force this sort of
fallback to a full-featured interpreter resolve when that is desired.

Fixes #1355