Un-reify installed wheel script shebangs. (#1410)
This fixes installed wheel chroots by re-writing Python script shebangs
to use the standard `#!python` placeholder which removes a
system-specific path from the installed wheel chroots shipped in PEX
files; leading to broader circumstances under which a PEX build is
reproducible. A set of failing test permutations are added to the
`test_reproducible_*` integration tests which this un-reification fixes.

Fixes #1407.