Fix `--venv` mode venv dir hash.
Previously the venv dir hash locked unconstrained PEXes into a venv using
the interpreter they were first run with. Now, when a `--venv` mode PEX
is either unconstrained or only constrained by a PEX_PYTHON_PATH that
admits the interpreter being used to execute the `--venv` mode PEX, that
interpreter is recorded in the venv dir hash to differentiate a venv
dedicated to that interpreter.

Fixes #1422