Clarify PEX_PYTHON & PEX_PYTHON_PATH interaction. (#1427)
Previously, requesting a PEX_PYTHON with an absolute path to an
interpreter might not fail if the interpreter was invalid or if it was
valid but not on a requested PEX_PYTHON_PATH. Fix up both behaviors and
close #431 as no longer relevant or well guided.

Clarifies behaviors #1422 should support in interpreter selection
effects on `--venv` mode venv directory hashes.