Enrich resolve and download with lock data. (#1437)
For now this data is only used in test_locked_resolve, but it forms the
basis for adding lock generation support to Pex.

This is the core of the work for #1401. The remaining work is wiring a
CLI option to generate a lock file (needs to be able to store multiple
platform specific locks since a single resolve can generate many) and
consume the lock file.