Remove zipapp execution mode & introduce --layout. (#1438)
The only remaining benefit of the traditional PEX execution mode over
the `--unzip` mode was that it was sensitive to saving space in the
PEX_ROOT with content-addressable caches of the PEX user code and third
party dependencies. With the introduction of the `--spread` layout,
which partially unpacked a spread PEX and was also sensitive to saving
space; using the same content-addressable caches, there was impetous to
re-work `--unzip` mode extraction to do said same.

Also, previously, a PEX could only be built via the CLI as a traditional
zipapp or in the new `--spread` directory layout. PEX has always
fundamentally produced a loose format via its `PEXBuilder.freeze` API
but this was never exposed as a CLI option.

Introduce a `--layout` CLI option to select between zipapp, packed (
formerly `--spread`) and loose PEX application layouts and unify
unpacking code, making the old `--unzip` execution mode the default and
consolidating on just that default and `--venv` for runtime execution
modes, neither of which suffer from `--not-zip-safe` issues. Now the
application developer is completely un-affected by the PEX runtime.
Loose files at development time are always loose files at runtime.

The choices that remain are:
1. Is my initial application startup time important to minimize?

   If so, the default PEX runtime execution mode is appropriate. An
   important example is a cloud lambda function. If not, use `--venv`
   mode for the minimal warm startup time and the best Python
   ecosystem compatibility.

2. How do I plan on delivering my PEX app?

   If the application will be delivered to a machine once, the default
   zipapp single file layout is very convenient. If the application will
   need to be incrementally updated, one of the directory modes is
   probably the best fit. The packed mode ensures you only ever sync a
   given version of the PEX runtime or a given version of a dependency
   once. The loose mode ensures you only every sync a given version of 
   a given file once, which allows for the maximum bandwidth savings.