Support `class Sub(Base[Concrete]): pass`. (#1457)
The prior implementation of `Generic` at runtime did not allow for this
and upcoming work on Pex CLI command infrastructure will want to do
this.

This new formulation needs a custom metaclass and that causes a problem
for `OrderedSet` since it already has an (indirect) custom metaclass of
ABCMeta via MutableSet. Re-work the typing there to accomodate.

Work towards #1401.