Introduce the `pex3 lock create` command. (#1464)
This provides a basic platform specific lock generation facility.

For now, the lock is only useful when a single platform or interpreter
is selected by the target options. A follow-up will introduce a lock
format that can store multiple locks allowing use and export of these
as individual units.

Work towards #1401.