Introduce a JSON codec for Lockfiles.
This will serve as an initial format for storing the public Lockfile
data and so must be robust and able to be evolved. A `pex_version` field
is included initially to provide the 1st hook for conditional handling
of the format as it evolves. It will probably make sense to introduce a
format version independent of the pex version, but this initial marker
allows that choice to be deferred.