Make resolution of DistributionTargets lazy.
Break up TargetConfiguration into just the configuration data and a
resolve method that can be used to request DistributionTargets only
when needed. This divide will be needed for platform agnostic locking
where interpreter constraints need to be known and used but
interpreters do not need to be resolved using those constraints.

Work towards #1402.