Never default marker environment entries.
We fail resolves that require unknown environment markers when building
a PEX via Pip and we should do the same when resolving from a PEX.
Although this case has not been knowingly exercised yet (Pants only
subsets for local interpreters and not foreign platforms), it is a both
inconsistent and incorrect behavior.

We'll have a similar need to resolve using platforms against lock files
shortly and all three forms of resolve shoule behave the same.