Introduce Rank and use it unify resolve ranking.
Previously some parts of the codebase used larger ints to denote better
fits and others used smaller ints for this. This resulted in a
mix of sorts of ranked things in the codebase which could be confusing.

Push the ordering, choosing and sorting into Rank to help make sense of
all this and ensure subset rwsolve ranking works the same as lockfile
resolve ranking.