Fix venv creation for duplicate symlinked dists. (#1639)
The fix in #1632 introduced a regression for the narrow case of `--venv`
mode PEXes using `--no-venv-site-packages-copies` (symlinks) and running
with a `PEX_PATH` including other PEXes with duplicate distributions
where those distributions provide only top-level modules (as opposed to
packages). Fix this issue in `Record.reinstall` and also fix duplicate
handling in `PEX.resolve` and `PEXEnvironment.activate`. Add tests that
cover all of this.

Fixes #1637