Go back to doing type conversion at caller
The type conversion is unfortunate that we go from an already parsed `pkg_resources.Requirement` to `str`, only to parse again. I don't want to hide that conversion behind the helper function `parse_lockable_requirements` - that should take the format it actually needs, and then callers do the inefficient conversion.