Switch Pex installed wheels to `--prefix` scheme.
Previously Pex used Pip's `--target` scheme which had both known bugs
(https://github.com/pypa/pip/issues/7658) and unknown quirks that Pex
was failing to fully be able to work around. Switch to the `--prefix`
scheme which exactly mirrors the scheme venvs use so that venvs can be
created with content of all sorts placed where it belongs.

This removes fragile parsing and interpretation of the RECORD; now Pex
only creates a RECORD, which is much more straight forward, when
building a venv.

Fixes #1656