Fix pathologic lock creation slowness.
For `universal` locks using the `pip-2020-resolver` and
`--interpreter-constraint` we could hit pathologically slow resolves
when there were many transitive requirements and / or many of these
were resolved as sdists. The issue was narrowed to `Requires-Python`
evaluation and this is fixed here to still be correct, but to do the
minimal amount of evaluation and backtracking.

Fixes #1706