Don't clear lock link database during analysis.
When we clear the database and a resolve back-tracks, we lose the
additional artifact urls needed to fill in the backtracked requirement
lock encountered later in the resolve.

Previously the clearing was done to relieve memory pressure during the
analysis but measurements show even moderately large resolves collect
only a handful of MBs worth of link data. We could collect 100x more
links in a very large resolve and still be OK here on all but the most
constrained machines. As such, we simply stop clearing the crawled link
database during analysis.

Fixes #1711