Factor Pip patching into DownloadObservers.
This moves patching out of the Pip tool into dedicated foreign platform
and locker download observers that do their own patching. This allows
the notion of locking to be narrowed to just the lock create use of Pip
download.

This is part of a series of a refactors needed to remove dependency
cycles in support of #1696.