Fix `pex3 lock export`: re-use `--lock` resolver. (#1831)
Previously the lock export utility was using out-dated logic from the
early days of the lock project. Extract a `subset` operation from the
lock resolver and have it and lock export share that code.

Also properly plumb lock export target selection and validation of
exactly one target with updated help and error messages to make it clear
that locks can only be exported for a single target.

Fixes #1826
Closes #1645