Fix `--venv` PEX `PEX_EXTRA_SYS_PATH` propagation.
Previously this used `PYTHONPATH` for propagation, which could lead to
issues executing foreign Pythons in subprocesses. Switch to using a
`.pth` file. Although the problem of polluting foreign Pythons in
subprocesses remains for zipapp PEXes, it is the case that `--venv` is
Pex's maximum compatibility mode and this can be viewed as yet another
case where a Pex feature breaking a Python app should be solved by using
`--venv` mode.

Fixes #1836