Add support for importing from PEXes. (#1845)
You can now use a PEX as a mostly normal `sys.path` entry. You do need
to activate it though and that can be done in two ways:

1. Perform an `import __pex__` before importing anything else you
   might need that lives in a PEX `sys.path` entry.
2. Import what you need in one step via `from __pex__ import colors` in
   the case where the PEX contains `ansicolors`, for example.

The second method is useful in cases where you can't influence which
python interpreter to run (In which case you could just specify the PEX
file itself), but you can influence the `sys.path`. The motivating
example of this is AWS Lambda functions for Python. One mode of
deploying these is in a zip. Here you can't pick the Python interpreter,
just its version, but you can specify the zip and an entry point. As
such you can specify a PEX as the zip and a `__pex__` psuedo-package
prefixed entry-point. This should obviate the need for projects like
Lambdex or at least simplify their integration with Pex, which currently
relies on using the unsupported `pex.pex_bootstrapper` APIs which I'd
like to drop in Pex 3.

Fixes #1839