Migrate PY27 pyenv to PY39.
This brings the 3 canned pyenv interpreters to 3.8, 3.9 and 3.10 paving
the way for pyenv-win. Although PY27 is dropped from pyenv supported
interpreters, new utilities are added to allow tests to depend on a
local pre-existing Python 2.7 install when present for tests that
intentionally exercise Python 2.7 and cannot be migrated to PY39.