Migrate PY27 pyenv to PY39. (#1914)
This brings the set of canned pyenv interpreters to 2.7 and 3.{8,9,10},
paving the way for pyenv-win. PY27 use is minimized and made explicit
and lazy such that appropriate `pytest.skip()`s can be added in the
Windows port to skip over the handful of ramining integration tests that
truly mean to test against a Python 2.7 interpreter.