Add a Zip utility for manipulating zip headers. (#1926)
This is not needed by Pex mainline but is needed to support Windows
console scripts which are realized as a sandwich of a windows PE stub,
a shebang line and a zipapp trailer. In this arrangement, the stub
executable header reads the zip trailer to find the sandwiched shebang
and then uses that shebang to pick a Python interpreter to run the
sandwich with. Pex needs to customize the sandwiched shebang just like
it needs to customize unix console script header shebangs so that the
venv interpreter is used to run the console script.