Don't treat `PEX_PATH=` as `.` like other PATHS.
This fixes a regression reading the PEX_PATH variable introduced by
changes in #1902 that ended up attempt to add CWD as a PEX when
PEX_PATH was set to the empty string in the environment.

Since PEX_PATH is closest in spirit to a PATH of files instead of a
PATH of directories like PATH and PYTHONPATH the old relied upon
behavior make sense generally.

Fixes #1936