Support injecting args and env vars in a PEX.
Pex now supports setting defaults for environment variables and
injecting default arguments for the entry point chosen at PEX build
time. This eliminates the need for projects like
https://github.com/kwlzn/pyuwsgi_pex and makes it easy to package a
PEXed framework app where the main entry point is fixed in the
framework and the hooks are configured to point to user supplied entry
points.

This removes the last known legitimate API use of Pex and paves the way
for Pex 3 to be cut in good conscience with a CLI-only API that does not
hinder any known use cases.

Fixes #987