Check for --complete-platforms match when --resolve-local-platforms (#1991)
This makes --resolve-local-platforms work better with
--complete-platforms: if there's an external complete platform specified
for a platform that matches a local interpreter, the local interpreter's
complete platform is checked against that external complete platform. If
the local interpreter supports any additional tags, the local
interpreter is not used.

If the local interpreter/environment is newer (supports more tags), the
old behaviour would risk including wheels that aren't compatible with
the external complete platform. That is, execution will fail, see
https://github.com/pantsbuild/pants/issues/17621 for an example.

Fixes #1899