Fix interpreter resolution when using --complete-platform with --resolve-local-platforms (#2031)
Fixes https://github.com/pantsbuild/pex/issues/2026.

Previously the interpreter resolution was more strict than necessary.
The `supported_tags` were only checked if the most specific tag of the
complete platform matched the platform of the local interpreter. This
prevented resolution in cases when it was possible and safe (ie all
`supported_tags` of the interpreter as present in the `supported_tags`
of the complete-platform).

Added an additional test for the above case. Fixed existing code and
also improved various existing tests by adding
`expected_complete_platforms`.