Handle file:// URL deps in distributions.
This has been a very long standing bug. The root issue is that `pip
download`, although it processes these deps and fails if not present on
the local filesystem, only handles resolving recursive dependenecies and
it does not "download" the local distribution from its referenced path
on the file system to the `pip download` download dir. As such, the
normal Pex post-processing is foiled for these deps.

The fix provided here is in common infra used by both normal Pex builds
and `--lock` builds and uses a recursive post-processing step that
gathers all such dependencies from the downloaded distributions and
injects them into the normal Pex post-processing chain.