Lock sdists with `prepare-metadata-for-build-wheel`.
PEP-517 specifies an optional `prepare-metadata-for-build-wheel` hook.
This can be used to extract metadata from an sdist without actually
building a wheel from the sdist when supported by the PEP-517 build
backend. Augment the Pex lock creation code to 1st try this and only
fall back to building a wheel when the optional hook is not present.

This allows locking sdist-only platform-specific projects that are
not compilable on the current platform by avoiding compiling them.

Fixes #2050