Attempt "cross-builds" of sdists for foreign platforms.
The "cross-building" is in scare-quotes because this is not actually
cross-building, it is just attempting a build of the sdist, and, if
successful, seeing if the resulting wheel matches the foreign platform.
This enables sdist-only projects to be used in foregin platform Pex
operations when it turns out the sdist is multi-platform.

Closes #2073