Eliminate warnings for default use. (#2188)
Previously `pex` would warn for default usage scenarios under both
Python 3.11 and Python 3.12. For Python 3.12, the `--pip-version`
defaulted to 23.2 but a warning about vendored Pip 20.3.4 was issued
anyway; this is now fixed. For both Python 3.11 and Python 3.12, the
vendored version of packaging used the deprecated `sre_constants`
stdlib module; this is now fixed with packaging 23.1 being vendored
and selected for Python >= 3.7.

Fixes #2186