Fix the Pex CLI to work when installed read-only.
When Pex is run from a read-only environment such as a venv that has
been made read-only after the fact or a nix environment, it must not
attempt to over-write its own vendored `.bootstrap` code in the process
of creating PEX runtime environments. Fix `.bootstrap` creation to only
write bootstrap files once.

Fixes #2203