Guard against too long filenames on eCryptFS.
The `AtomicDirectory` mechanism inflates the target directory file name
length. In the case of unlocked operation in particular, this can lead
to file name length violations when the PEX_ROOT is the default `~/.pex`
and the home dir is encrypted with eCryptFS, which generally limits
file name lengths to 143 characters. Guard against this explicitly.

Fixes #2087