Add `--devpi` support to tox test environments. (#2198)
The new `--devpi` support is opt-in for now and CI opts-in. The speed 
gains are modest and variable in CI where the network connection has
low latency and high bandwidth, but this also should improve stability 
there with less network flakes (anecdotally true from a few weeks of 
using this locally and retrying in CI). It's also the right thing to do
vs. "hammering" PyPI for no reason.