Traverse directories in stable order when building a PEX (#2220)
Addresses one part of #2155.

Adds a wrapper for `os.walk` that sorts directory entries as it
traverses directories. As discussed in the issue, `os.walk` may yield
directory entries in a different order depending on (possibly) the file
system.

I don't know if the two `os.walk` calls that this PR replaces are enough
for addressing this class of non-determinism in pex, but these were
sufficient for solving the issue in my case. Happy to expand the use of
the wrapper if there are other places that it should be used.