Fix `.readthedocs.{yml,yaml}` duplication. (#2275)
The fix in #2274 that introduced `.readthedocs.yaml` (the name
documented by RTD currently to use for the config file) was blind to the
fact we already had a `.readthedocs.yml` file. It appears to be the case
that `.readthedocs.yml` is both read by RTD and preferred; so RTD build
errors persisted. Fix this by merging the config files and sticking with
the current reccomended name of `.readthedocs.yaml`.

More work towards #2267.