Transition from Installed to ResolveResult.
This change renames types and variables to take the emphasis off
installation and put it on the more generic result of a resolve (which
will soon include raw `.whl` files as an option in addition to default
installed wheel chroots).

Also factor out the machinery needed to be able to ask `BuildAndInstall`
for `.whl` file results in addition to the current installed wheel
chroot results.