Trim embedded doc image assets.
Re-use the Pex icon png for the favicon and delete `pex.ico`. Browser
support for png favicons is both wide and deep at this point and
removing the ~46KB `.ico` file trims ~20KB from the binary distributions
we release.

Also re-name pex-full* assets to pex-logo* to reflect their role and
tidy up the SVG source. There is the SVG source and then the logo and
icon derived from it. This should be a bit more clear now.