Implement pex3 lock sync.
Introduce the high level `pex3 lock sync` command. This should generally
suffice for typical use cases and works as follows:

+ On first use (where the specified `--lock` does not yet exist), it
  acts just like `pex3 lock create`.
+ On subsequent uses it does a minimal synchronization of the lock based
  on the diff of the given current requirements against the requirements
  used to generate the specified `--lock`. This amounts to formulating a
  `pex3 lock update` command with the appropriate `-p`, `-R` and `-d`
  arguments.

In addition to creating and syncing a lock, it can also create and sync
a venv (--venv) based on the lock. Further, a command can be specified
to run in the synchronized venv with arguments following the `--` option
terminator.

This latter set of features allow Pex to act as a concise tool in `tox` /
`nox` / `invoke` / `make` setups to implement a simple build system.

Fixes #2344